+++ demo/src/main/java/com/orange/ods/demo/ui/components/ComponentsNavigationItem.kt
@@ -24,6 +24,7 @@ sealed class ComponentsNavigationItem(
     object Cards : ComponentsNavigationItem(R.string.component_cards, "components/card", R.drawable.picture_component_cards)
     object Controls : ComponentsNavigationItem(R.string.component_controls, "components/controls", R.drawable.picture_component_controls)
     object Progress : ComponentsNavigationItem(R.string.component_progress, "components/progress", R.drawable.picture_component_progress)
+    object Lists : ComponentsNavigationItem(R.string.component_lists, "components/lists", R.drawable.picture_component_lists)
 }
 
 sealed class ComponentsSubLevelNavigationItem(
@@ -38,4 +39,8 @@ sealed class ComponentsSubLevelNavigationItem(
     object ControlsRadioButtons : ComponentsSubLevelNavigationItem(R.string.component_controls_radio_buttons, "components/controls/radio_buttons")
     object ControlsSwitches : ComponentsSubLevelNavigationItem(R.string.component_controls_switches, "components/controls/switches")
     object ControlsSliders : ComponentsSubLevelNavigationItem(R.string.component_controls_sliders, "components/controls/sliders")
+
+    object ListsOneLine : ComponentsSubLevelNavigationItem(R.string.component_lists_one_line, "components/lists/one_line")
+    object ListsTwoLines : ComponentsSubLevelNavigationItem(R.string.component_lists_two_lines, "components/lists/two_lines")
+    object ListsThreeLines : ComponentsSubLevelNavigationItem(R.string.component_lists_three_lines, "components/lists/three_lines")
 }