+++ demo/src/main/java/com/orange/ods/demo/ui/utilities/MenuItem.kt
@@ -1,0 +1,15 @@
+/*
+ *
+ *  Copyright 2021 Orange
+ *
+ *  Use of this source code is governed by an MIT-style
+ *  license that can be found in the LICENSE file or at
+ *  https://opensource.org/licenses/MIT.
+ * /
+ */
+
+package com.orange.ods.demo.ui.utilities
+
+import androidx.annotation.StringRes
+
+data class MenuItem(@StringRes val titleRes: Int, val route: String)